requirements.txt
streamlit==1.0.0
psycopg2==2.9.1
pandas==1.3.3
plotly==5.3.1
python-dotenv==0.19.0
requests==2.26.0
